Add AOSP build system support
Add the Makefile and remove the dependency on Google Play Services
when building TalkBack from an AOSP tree.

The Google Play Services library is needed for a runtime detection
of Google Play Services. If available, TalkBack will add a link
to Google TalkBack on the Play Store. Since the linked app is not
compatible with the one built, add a set of classes that replace
the ones imported from the Google Play Services library and use
them to mimic the lack of Play Services on the device. This ensures
that the link is never aded and allows to not depend on the Google
Play Services library to build TalkBack.

In addition to that, create a replacement for gradle's BuildConfig
and replace ${applicationId} at build time.

This commit doesn't affect the builds done with gradle, which will
keep working as before and will still depend on the Google Play
Services library. The test package can still be built only using
gradle.